requirements.txt
lxml==4.3.4
requests==2.22.0
tqdm==4.32.2
kafka-python==1.4.6
